pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "behapy"
version = "0.4.0"
description = "A Scanpy-like framework for analyzing high-dimensional behavioral time series data"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Wulin Tan", email = "wulin.tan9527@gmail.com"}
]
maintainers = [
    {name = "Wulin Tan", email = "wulin.tan9527@gmail.com"}
]
dependencies = [
    "anndata>=0.8.0",
    "scanpy>=1.9.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "umap-learn>=0.5.0",
    "numba>=0.54.0",
    "h5py>=3.0.0",
    "tqdm>=4.62.0",
    "leidenalg",
    "python-louvain",
    "hdbscan",
    "igraph",
    "networkx",
    "annoy>=1.17.0",
    "statsmodels",
]

[project.urls]
Homepage = "https://github.com/Wulin-Tan/behapy"
Repository = "https://github.com/Wulin-Tan/behapy"
Documentation = "https://github.com/Wulin-Tan/behapy/tree/main/docs"

[project.optional-dependencies]
gpu = ["cupy"]
video = ["opencv-python", "moviepy"]
interactive = ["plotly", "streamlit"]
dev = ["black", "isort", "pytest", "mypy", "flake8"]
docs = ["sphinx", "sphinx-rtd-theme", "nbsphinx"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
